third_party/blink/web_tests/external/wpt/performance-timeline/not-restored-reasons/performance-navigation-timing-fetch.tentative.window-expected.txt

This is a testharness.js-based test.
FAIL Ensure that ongoing fetch upon entering bfcache blocks bfcache and recorded. assert_equals: document unexpectedly BFCached expected (undefined) undefined but got (boolean) true
Harness: the test ran to completion.
